% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSDSDemographicGrowthApp-package.R
\docType{package}
\name{CSDSDemographicGrowthApp-package}
\alias{CSDSDemographicGrowthApp}
\alias{CSDSDemographicGrowthApp-package}
\title{CSDSDemographicGrowthApp: Exploring the impact of demographic growth on community services demand}
\description{
A Shiny application to show the projected impact of demographic growth on community services, using the NHS England CSDS dataset.
}
\author{
\strong{Maintainer}: Francis Barton \email{francis.barton@nhs.net}

Authors:
\itemize{
  \item Rhian Davies \email{rhian.davies25@nhs.net}
}

Other contributors:
\itemize{
  \item The Strategy Unit [funder]
}

}
\keyword{internal}
