% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_projected_contacts_fy}
\alias{nat_projected_contacts_fy}
\title{Projected national contact numbers by service and financial year}
\format{
\subsection{nat_projected_contacts_fy}{

A tibble with 1 row and 10 columns, including a list-column with a nested
tibble, which has 6 columns and 101,283 rows.
\describe{
\item{nat_contacts_all_unfiltd}{The total number of contacts in CSDS, nationally - integer}
\item{nat_contacts_missing_icb}{The number of contacts not allocated to an ICB, nationally - integer}
\item{nat_contacts_missing_gen}{The number of contacts with missing or invalid patient gender category, nationally - integer}
\item{nat_contacts_missing_age}{The number of contacts with no patient age recorded, nationally - integer}
\item{nat_contacts_inconsistnt}{The number of contacts included in submissions by inconsistent submitters, nationally - integer}
\item{nat_contacts_not_attnded}{The number of contacts where the appointment was not attended, nationally - integer}
\item{nat_contacts_canclld_unk}{The number of contacts where the appointment was cancelled, or the appointment status was unknown, nationally - integer}
\item{nat_contacts_total_excld}{The total number of contacts to be excluded by any of the criteria above, nationally - integer}
\item{nat_contacts_final_count}{The total number of remaining contacts to be included in the tool, nationally - integer}
\item{data}{List of tibbles containing projected contacts by financial year and single year of age, nationally - list}
\item{fin_year}{Financial year (in nested tibble) - character}
\item{age_int}{Patient age at time of contact (in nested tibble) - integer}
\item{team_type}{The name of the service providing the contact (in nested tibble) - character}
\item{projected_contacts}{Number of projected contacts for each combination of financial year, age, and team type, nationally (in nested tibble) - double}
\item{proj_popn_by_fy_age}{Projected population for each combination of financial year and age, nationally (in nested tibble) - double}
\item{projected_contacts}{Number of projected contacts for each combination of financial year and age, nationally (in nested tibble) - double}
}
}
}
\usage{
nat_projected_contacts_fy
}
\description{
CSDS data is combined with ONS sub-national population projections (by LAD)
for the period 2022-2043. The population projections are based on 2018
population estimates. These are recalculated to create population
projections for each financial year from 2022/23 to 2042/43.
}
\keyword{datasets}
